Break up builder into separate files
