+++ src/marshmallow_annotations/exceptions.py
@@ -1,0 +1,10 @@
+from marshmallow.exceptions import MarshmallowError
+
+
+class MarshmallowAnnotationError(MarshmallowError):
+    pass
+
+
+class AnnotationConversionError(MarshmallowAnnotationError):
+    pass
+
